requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Task queue
celery==5.4.0
redis==5.1.0

# Document parsing
pymupdf==1.24.10
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5

# LLM
anthropic==0.34.0

# Utilities
python-multipart==0.0.9
httpx==0.27.2
aiofiles==24.1.0
python-dateutil==2.9.0
jinja2==3.1.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
